src/backend/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.3
pdfplumber==0.11.0
pdfminer.six==20231228
python-docx==1.1.0
pandas==2.2.1
